requirements.txt
asgiref==3.9.1
boto3==1.39.7
botocore==1.39.7
Django==5.2.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
django-storages==1.14.6
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
dj-database-url==3.0.1
whitenoise==6.9.0
setuptools==80.9.0
pillow==11.3.0
pyjwt==2.10.1
python-decouple==3.8
stripe>=13.0.0
paypal-server-sdk==1.1.0
django-environ>=0.10.0  # for secure env management






